forgot password functionality moved to forgotRoute folder
